hardware/raspberry_pi/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
RPi.GPIO>=0.7.0
rplidar>=1.0.0
scipy>=1.7.0
pyserial>=3.5
smbus2>=0.4.0